+++ frontend/src/queries/templates/useTemplateExploreQuery.ts
@@ -1,6 +1,7 @@
 import { keepPreviousData, useQuery } from '@tanstack/react-query';
 
-import { ApiError, PAGE_SIZE, QUERY_KEY, DEFAULT_SORTING_OPTION, getTemplateExplore } from '@/api';
+import { PAGE_SIZE, QUERY_KEY, DEFAULT_SORTING_OPTION, getTemplateExplore } from '@/api';
+import { ApiError } from '@/api/Error';
 import type { SortingKey, TemplateListResponse } from '@/types';
 
 interface Props {
